fix: new signal proxy
