+++ projects/ngrx-rtk-query/src/lib/utils/index.ts
@@ -1,4 +1,4 @@
 export * from './capitalize';
-export * from './deep-signal';
 export * from './shallow-equal';
+export * from './signal-proxy';
 export * from './tsHelpers';
